requirements.txt
cmake==3.22.4
flake8==4.0.1
Flask==2.0.2
Flask-Cors==3.0.10
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
Pillow==9.0.1
mysql-connector-python==8.0.27
mysql-connector-python-rf==2.2.2
face-recognition==1.3.0
face-recognition-models==0.3.0